+++ openghg/types/_enum.py
@@ -79,3 +79,7 @@ class DataTypes(Enum):
     NPL = "NPL"
     OPENGHG = "OPENGHG"
     INTEM = "INTEM"
+
+
+class BoundaryConditions(Enum):
+    OPENGHG = "OPENGHG"